test: extract bulk stubs into their own files
